% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_growth_rates.R
\name{empirical_growth_rates}
\alias{empirical_growth_rates}
\title{Title}
\usage{
empirical_growth_rates(
  count_data,
  time_data,
  num_train,
  num_test,
  population,
  susc_perc,
  by_factor = NULL
)
}
\arguments{
\item{count_data}{Response vector of Count Data}

\item{time_data}{Covariate vector of Time Data}

\item{num_train}{Constant. Represents the number of training data points.}

\item{num_test}{Constant. Represents the number of testing data points. Use 0 test days for a full empirical growth rate model.}

\item{population}{Constant. Represents the total population.}

\item{susc_perc}{Constant between 0 and 1. Represents the percent of the population that is susceptible}

\item{by_factor}{(Optional) Additional categorical covariate}
}
\value{
A list containing
\item{data}{A dataframe containing time series data empirical growth rate estimates}
\item{emp_grow_mod}{Model used to calculate empirical growth rate estimates}
}
\description{
Title
}
\examples{
# Example dataset
data("jh_data_daily_confirm")
data("state_population")

# Specify the count data to be days 200-400 of Florida's daily Covid-19 cases
count_data = jh_data_daily_confirm$Florida[200:400]

# Specify the time data to be days 200-400
time_data = jh_data_daily_confirm$date[200:400]

# Add a "day of week" effect to the model
by_factor = jh_data_daily_confirm$day_of_week[200:400]

num_train = 28
num_test = 0

# Get Florida Population and use .55 as susceptible population proportion
population = state_population$Florida[1]
susc_perc = .55

Florida_emp_grow_rate = empirical_growth_rates(count_data = count_data,
                                               time_data = time_data,
                                               num_train = num_train,
                                               num_test = num_test,
                                               population = population,
                                               susc_perc = susc_perc,
                                               by_factor = by_factor)
}
