% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_negative_binomial.R
\name{fit_negative_binomial}
\alias{fit_negative_binomial}
\title{fit_negative_binomial}
\usage{
fit_negative_binomial(
  count_data,
  time_data,
  by_factor = NULL,
  cooksConstant = 4,
  plot_model = TRUE
)
}
\arguments{
\item{count_data}{Response vector of count data to be modeled}

\item{time_data}{Covariate of time data used in the model}

\item{by_factor}{(Optional) Additional categorical covariate}

\item{cooksConstant}{(Optional) Specify the sensitivity to outliers. Lower values give higher sensitivity.}

\item{plot_model}{(Optional) Boolean argument to plot the fitted values of the model}
}
\value{
A list containing
\item{model}{mgcv::gam object. A spline model of count_data response variable and time_data covariate with by_factor as a possible factor}
\item{data}{Dataframe object. A data frame with binded count_data, time_data, by_factor (if used), and fitted value from model}
\item{cooks_distances}{Calculated cooks distances from model}
\item{outliers}{Flagged outliers based on cooks distances and cooksConstant}
}
\description{
fit_negative_binomial
}
