% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_poisson.R
\name{fit_poisson}
\alias{fit_poisson}
\title{fit_poisson}
\usage{
fit_poisson(
  count_data,
  time_data,
  by_factor = NULL,
  cooksConstant = 4,
  return_plot = FALSE
)
}
\arguments{
\item{count_data}{Response vector of count data to be modeled}

\item{time_data}{Covariate of time data used in the model}

\item{by_factor}{(Optional) Additional categorical covariate}

\item{cooksConstant}{(Optional) Specify the sensitivity to outliers. Lower values give higher sensitivity.}

\item{return_plot}{(Optional) Boolean argument to return a ggplot2 plot object}
}
\value{
A list containing
\item{model}{mgcv::gam object. A spline model of count_data response variable and time_data covariate with by_factor as a possible factor}
\item{data}{Dataframe object. A data frame with binded count_data, time_data, by_factor (if used), and fitted value from model}
\item{cooks_distances}{Calculated cooks distances from model}
\item{outliers}{Flagged outliers based on cooks distances and cooksConstant}
\item{plot}{ggplot2 plot object of model fitted values}
}
\description{
fit_poisson
}
\examples{

# Example dataset
data("jh_data_daily_confirm")

# Specify the count data to be the first 275 days of Arizona's daily Covid-19 cases
count_data = jh_data_daily_confirm$Arizona[1:275]

# Specify the time data to be the first 275 days
time_data = jh_data_daily_confirm$date[1:275]

# Add a "day of week" effect to the model
by_factor = jh_data_daily_confirm$day_of_week[1:275]


Arizona_poisson = fit_poisson(count_data, time_data, by_factor, return_plot = FALSE)

}
