% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_screenshot.R
\name{insert_screenshot}
\alias{insert_screenshot}
\title{Salva uma tabela gerada com o pacote flextable,
exportando como uma figura png}
\usage{
insert_screenshot(x, png_file = "image.png")
}
\arguments{
\item{x}{tabela gerada pelo pacote flextable}

\item{png_file}{nome que o arquivo sera salvo}
}
\description{
A partir de uma tabela html gerada com o pacote
flextable tira-se um webshot da tabela html para
salvar a imagem como uma figura png.
}
\details{
O link usado de referencia para criar a funcao e
\url{https://stackoverflow.com/a/49049701/9990992}
}
\examples{
flextable(mtcars) \%>\% insert_screenshot()
flextable(mtcars) \%>\% insert_screenshot('my-screenshot.png')

}
