% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_t.R
\name{lambda_t}
\alias{lambda_t}
\title{lambda_t function of equation 17}
\usage{
lambda_t(k, phi)
}
\arguments{
\item{k}{A integer number between "last day - num_test day + 1" and "last day + 1"}

\item{phi}{A double greater than 0}
}
\value{
scalar defined by equation 17
}
\description{
lambda_t function of equation 17
}
\examples{

last_day = 497
num_test_days = 14

lambda_t(490, 0.7)
}
