% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_mexico_random_vectors_example}
\alias{new_mexico_random_vectors_example}
\title{New Mexico Random Vector Example}
\format{
\subsection{\code{new_mexico_random_vectors_example}}{

A data frame with 100 rows and 3 columns:
\describe{
\item{eta}{100 random eta variables. Used in coffee_forecast/coffee_method.}
\item{omega}{100 random omega variables. Used in coffee_forecast/coffee_method.}
\item{phi}{100 random phi variables. USed in coffee_forecast/coffee_method.}
...
}
}
}
\usage{
new_mexico_random_vectors_example
}
\description{
A set of 100 random vectors drawn from the joint pdf defined by the inverse distance function.
Uses empirical growth data generated from the first 273 days of New Mexico Covid-19 data.
}
\keyword{datasets}
