% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_t.R
\name{omega_t}
\alias{omega_t}
\title{omega_t function of equation 16}
\usage{
omega_t(k, omega)
}
\arguments{
\item{k}{A integer number between "last day - num_test day + 1" and "last day + 1"}

\item{omega}{A double greater than or equal to 1}
}
\value{
scalar defined by equation 16
}
\description{
omega_t function of equation 16
}
\examples{

last_day = 497
num_test_days = 14

omega_t(490, 1.2)
}
