% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_set.R
\name{palette_set}
\alias{palette_set}
\title{Color Palette - Paleta de Cores}
\usage{
palette_set(paleta = NULL)
}
\description{
Essa funcao retorna a alguma das seguintes paletas:
\itemize{
  \item \strong{prim}: Azul Escuro, Azul Claro
  \item \strong{prim2}: Azul Marinho, Azul Escuro, Azul Medio, Azul Claro, Azul Celeste
  \item \strong{sec}: Cinza
  \item \strong{sec2}: Tons de Cinzas (mais escuro para o mais claro)
  \item \strong{aux}: Amarelo, Vinho, Laranja, Roxo, Turquesa, Oliva
  \item \strong{aux2}: Amarelo Claro, Amarelo Escuro, Laranja Claro, Laranja Escuro, Vinho Claro, Vinho Escuro, Roxo Claro, Roxo Escuro, Turquesa Claro, Turquesa Escuro, Oliva Claro, Oliva Escuro
  \item \strong{vibrante}: Roxo Vibrante, Roxo Vibrante Claro, Roxo Vibrante Escuro, Azul Vibrante, Azul Vibrante Claro, Azul Vibrante Escuro, Verde Vibrante, Verde Vibrante Claro, Verde Vibrante Escuro
}
}
\keyword{color}
\keyword{palette}
