% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_screenshot.R
\name{save_screenshot}
\alias{save_screenshot}
\title{Salva uma tabela gerada com o pacote flextable,
exportando como uma figura png}
\usage{
save_screenshot(filename, file)
}
\arguments{
\item{filename}{Nome do arquivo; O qual um objeto foi atribuido.}

\item{file}{Nome do arquivo + .png ou .jpg}
}
\description{
O link usado de referencia para criar a funcao e
\url{https://stackoverflow.com/questions/50225669/how-to-save-flextable-as-png-in-r}
}
\details{
The param must be written between " "
O parametro Deve ser escrito entre " "
}
\examples{
arquivo <- head(iris) \%>\% regulartable()
save_screenshot("arquivo", "C:/Users/fulano/Desktop/arquivo.png")

}
