% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_population}
\alias{state_population}
\title{State Population - United States}
\format{
\subsection{\code{state_population}}{

A data frame with 4 rows and 54 columns:
\describe{
\item{state/territory}{Population of State/Territory}
\item{US}{Population of US}
\item{year}{Year in numeric format}
...
}
}
}
\source{
\url{https://www2.census.gov/programs-surveys/popest/tables/2020-2023/state/totals/NST-EST2023-POP.xlsx}
}
\usage{
state_population
}
\description{
A processed data set of US Population by State from The United States Census Bureau
Separated by state and Year 2020-203
}
\keyword{datasets}
